Update README.md
The sample size of the survey was 262 respondents which was randomly selected. The male respondents made 58% of the total sample size while 42% were females.
 
The age distribution is highly skewed to the left (between 18 and 40 years) and 87% of the customers are working class with majority in the private sector.
 
At least 30% of the respondents have the propensity to stop using the bank due to poor customer service, bad processes and security
concerns in that order.
 
Frequent cash out at the ATMs does not really influence the decision of respondents to/not patronize the bank.
 
Recommendations
Expand self-service to improve customer service
A real-time monitoring system can be a valuable tool
in improving service.